+++ fitsio/fitsio_pywrap.c
@@ -34,9 +34,16 @@
 // not sure where this is defined in numpy...
 #define NUMPY_MAX_DIMS 32
 
+// max len of python error message
+#define PYFITS_ERRMSG_LEN 1024
+
 struct PyFITSObject {
     PyObject_HEAD
     fitsfile* fits;
+    // we store the python error message here so that we record all error
+    // messages as they happen. sometimes cfitsio will clear
+    // the error stack and this removes important debugging info
+    char pyfits_errmsg[PYFITS_ERRMSG_LEN];
 };
 
 #ifdef FITSIO_PYWRAP_ALWAYS_NONSTANDARD_STRINGS
@@ -228,26 +235,60 @@ static char* get_object_as_string(PyObject* obj)
     return strdata;
 }
 
-static void set_ioerr_string_from_status(int status) {
+static void set_ioerr_string_from_status(int status, struct PyFITSObject* self) {
     char status_str[FLEN_STATUS], errmsg[FLEN_ERRMSG];
-    char message[1024];
+    char str_with_newline[1024];
+    int nleft;
+    // sometimes this function is called without an instantiated PyFITSPObject.
+    // in that case, we use a local buffer and grab all of the information we can
+    // to set the error string. this edge case means we will sometimes miss error
+    // messages that cfitsio puts on the stack but then removes later.
+    char message_if_null[PYFITS_ERRMSG_LEN];
 
-    int nleft=1024;
+    // pointer to the error message buffer we are using
+    char* message;
+
+    // always init error message to zero length string
+    // init of self->pyfits_errmsg is done when object is created
+    message_if_null[0] = '\0';
+
+    if (self != NULL) {
+        message = &(self->pyfits_errmsg[0]);
+    } else {
+        message = &(message_if_null[0]);
+    }
+
+    // the number of characters left is the number of characters not
+    // used in the message variable minus one. We remove one for the
+    // C null termination character. So if the variable is PYFITS_ERRMSG_LEN
+    // characters long, then nleft is PYFITS_ERRMSG_LEN - 1.
+    // This works because:
+    //   - strncat always adds the null termination character
+    //   - strlen returns the number of characters up to the first null (and
+    //     does not include the null itself in the total count).
+    // Thus we never have to account for the null ourselves except in the initial
+    // number of characters we have to use for holding error messages (which we
+    // set to one less than the total storage we have so strncat always has
+    // room for the null.)
+    nleft = PYFITS_ERRMSG_LEN - strlen(message) - 1;
 
     if (status) {
         fits_get_errstatus(status, status_str);  /* get the error description */
 
-        sprintf(message, "FITSIO status = %d: %s\n", status, status_str);
-
-        nleft -= strlen(status_str)+1;
+        sprintf(str_with_newline, "FITSIO status = %d: %s\n", status, status_str);
+        if (nleft >= strlen(str_with_newline)) {
+            strncat(message, str_with_newline, nleft);
+            nleft -= strlen(str_with_newline);
+        }
 
         while ( nleft > 0 && fits_read_errmsg(errmsg) )  { /* get error stack messages */
-            strncat(message, errmsg, nleft-1);
-            nleft -= strlen(errmsg)+1;
-            if (nleft >= 2) {
-                strncat(message, "\n", nleft-1);
+            sprintf(str_with_newline, "%s\n", errmsg);
+            if (nleft >= strlen(str_with_newline)) {
+                strncat(message, str_with_newline, nleft);
+                nleft -= strlen(str_with_newline);
+            } else {
+                break;
             }
-            nleft-=2;
         }
         PyErr_SetString(PyExc_IOError, message);
     }
@@ -440,6 +481,9 @@ PyFITSObject_init(struct PyFITSObject* self, PyObject *args, PyObject *kwds)
     int status=0;
     int create=0;
 
+    // init the error message to an empty string
+    self->pyfits_errmsg[0] = '\0';
+
     if (!PyArg_ParseTuple(args, (char*)"sii", &filename, &mode, &create)) {
         return -1;
     }
@@ -447,12 +491,12 @@ PyFITSObject_init(struct PyFITSObject* self, PyObject *args, PyObject *kwds)
     if (create) {
         // create and open
         if (fits_create_file(&self->fits, filename, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return -1;
         }
     } else {
         if (fits_open_file(&self->fits, filename, mode, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return -1;
         }
     }
@@ -470,7 +514,7 @@ PyFITSObject_repr(struct PyFITSObject* self) {
         char repr[2056];
 
         if (fits_file_name(self->fits, filename, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
 
@@ -489,7 +533,7 @@ PyFITSObject_filename(struct PyFITSObject* self) {
         char filename[FLEN_FILENAME];
         PyObject* fnameObj=NULL;
         if (fits_file_name(self->fits, filename, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
 
@@ -510,7 +554,7 @@ PyFITSObject_close(struct PyFITSObject* self)
     if (fits_close_file(self->fits, &status)) {
         self->fits=NULL;
         /*
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
         */
     }
@@ -601,7 +645,7 @@ PyFITSObject_movnam_hdu(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_movnam_hdu(self->fits, hdutype, extname,  extver, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -627,7 +671,7 @@ PyFITSObject_movabs_hdu(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     hdutypeObj = PyLong_FromLong((long)hdutype);
@@ -659,7 +703,7 @@ PyFITSObject_get_hdu_info(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -954,7 +998,7 @@ PyFITSObject_get_hdu_name_version(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -1295,7 +1339,7 @@ static int create_empty_hdu(struct PyFITSObject* self)
     int naxis=0;
     long* naxes=NULL;
     if (fits_create_img(self->fits, bitpix, naxis, naxes, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return 1;
     }
 
@@ -1312,7 +1356,8 @@ static int create_empty_hdu(struct PyFITSObject* self)
 //
 // note tile dims are written reverse order since
 // python orders C and fits orders Fortran
-static int set_compression(fitsfile *fits,
+static int set_compression(struct PyFITSObject* self,
+                           fitsfile *fits,
                            int comptype,
                            PyObject* tile_dims_obj,
                            int *status) {
@@ -1323,7 +1368,7 @@ static int set_compression(fitsfile *fits,
 
     // can be NOCOMPRESS (0)
     if (fits_set_compression_type(fits, comptype, status)) {
-        set_ioerr_string_from_status(*status);
+        set_ioerr_string_from_status(*status, self);
         goto _set_compression_bail;
         return 1;
     }
@@ -1414,7 +1459,7 @@ PyFITSObject_create_image_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
     // We do allow overriding the dither seed and then putting it back
     // Luckily cfitsio has public APIs for that so this should be robust
     if (fits_get_dither_seed(self->fits, &orig_dither_seed, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         goto create_image_hdu_cleanup;
     } else {
         // tell the code to put back the dither seed later
@@ -1526,28 +1571,28 @@ PyFITSObject_create_image_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
 
         if ((comptype_obj != Py_None) || (tile_dims_obj != Py_None)) {
             // exception strings are set internally
-            if (set_compression(self->fits, comptype, tile_dims_obj, &status)) {
+            if (set_compression(self, self->fits, comptype, tile_dims_obj, &status)) {
                 goto create_image_hdu_cleanup;
             }
         }
 
         if (qlevel_obj != Py_None) {
             if (fits_set_quantize_level(self->fits, qlevel, &status)) {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, self);
                 goto create_image_hdu_cleanup;
             }
         }
 
         if (qmethod_obj != Py_None) {
             if (fits_set_quantize_method(self->fits, qmethod, &status)) {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, self);
                 goto create_image_hdu_cleanup;
             }
         }
 
         if (dither_seed_obj != Py_None) {
             if (fits_set_dither_seed(self->fits, dither_seed, &status)) {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, self);
                 goto create_image_hdu_cleanup;
             }
         }
@@ -1555,21 +1600,21 @@ PyFITSObject_create_image_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
         if (comptype == HCOMPRESS_1) {
             if (hcomp_scale_obj != Py_None) {
                 if (fits_set_hcomp_scale(self->fits, hcomp_scale, &status)) {
-                    set_ioerr_string_from_status(status);
+                    set_ioerr_string_from_status(status, self);
                     goto create_image_hdu_cleanup;
                 }
             }
 
             if (hcomp_smooth_obj != Py_None) {
                 if (fits_set_hcomp_smooth(self->fits, hcomp_smooth, &status)) {
-                    set_ioerr_string_from_status(status);
+                    set_ioerr_string_from_status(status, self);
                     goto create_image_hdu_cleanup;
                 }
             }
         }
 
         if (fits_create_img(self->fits, image_datatype, ndims, dims, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             goto create_image_hdu_cleanup;
         }
     }
@@ -1578,12 +1623,12 @@ PyFITSObject_create_image_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
 
             // comments are NULL
             if (fits_update_key_str(self->fits, "EXTNAME", extname, NULL, &status)) {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, self);
                 goto create_image_hdu_cleanup;
             }
             if (extver > 0) {
                 if (fits_update_key_lng(self->fits, "EXTVER", (LONGLONG) extver, NULL, &status)) {
-                    set_ioerr_string_from_status(status);
+                    set_ioerr_string_from_status(status, self);
                     goto create_image_hdu_cleanup;
                 }
             }
@@ -1592,7 +1637,7 @@ PyFITSObject_create_image_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
 
     if (nkeys > 0) {
         if (fits_set_hdrsize(self->fits, nkeys, &status) ) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             goto create_image_hdu_cleanup;
         }
     }
@@ -1604,14 +1649,14 @@ PyFITSObject_create_image_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
         nelements = PyArray_SIZE(array);
         data = PyArray_DATA(array);
         if (fits_write_img(self->fits, datatype, firstpixel, nelements, data, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             goto create_image_hdu_cleanup;
         }
     }
 
     // this flushes all buffers
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         goto create_image_hdu_cleanup;
     }
 
@@ -1626,14 +1671,14 @@ create_image_hdu_cleanup:
         || (hcomp_smooth_obj != Py_None)
     ) {
         if (fits_unset_compression_request(self->fits, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
         }
     }
 
     // put back the dither seed after we reset any compression params
     if (got_dither_seed == 1) {
         if (fits_set_dither_seed(self->fits, orig_dither_seed, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
         }
     }
 
@@ -1672,13 +1717,13 @@ PyFITSObject_reshape_image(struct PyFITSObject* self, PyObject* args) {
     dims_array = (PyArrayObject *) dims_obj;
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // existing image params, just to get bitpix
     if (fits_get_img_paramll(self->fits, maxdim, &bitpix, &ndims_orig, dims_orig, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -1689,7 +1734,7 @@ PyFITSObject_reshape_image(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_resize_imgll(self->fits, bitpix, ndims, dims, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -1725,7 +1770,7 @@ PyFITSObject_write_image(struct PyFITSObject* self, PyObject* args) {
     array = (PyArrayObject *) array_obj;
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -1744,13 +1789,13 @@ PyFITSObject_write_image(struct PyFITSObject* self, PyObject* args) {
     nelements = PyArray_SIZE(array);
     firstpixel = (LONGLONG) firstpixel_py;
     if (fits_write_img(self->fits, datatype, firstpixel, nelements, data, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this is a full file close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -1856,7 +1901,7 @@ PyFITSObject_write_subset(struct PyFITSObject* self, PyObject* args) {
  * The keys are written as TDIM{colnum}
  */
 static int
-add_tdims_from_listobj(fitsfile* fits, PyObject* tdimObj, int ncols) {
+add_tdims_from_listobj(struct PyFITSObject* self, fitsfile* fits, PyObject* tdimObj, int ncols) {
     int status=0, i=0;
     size_t size=0;
     char keyname[20];
@@ -1896,7 +1941,7 @@ add_tdims_from_listobj(fitsfile* fits, PyObject* tdimObj, int ncols) {
             free(tdim);
 
             if (status) {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, self);
                 return 1;
             }
         }
@@ -1966,11 +2011,11 @@ PyFITSObject_create_table_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
     nfields = ttyp->size;
     if ( fits_create_tbl(self->fits, table_type, nrows, nfields,
                          ttyp->data, tform->data, tunit->data, extname_use, &status) ) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         goto create_table_cleanup;
     }
 
-    if (add_tdims_from_listobj(self->fits, tdimObj, nfields)) {
+    if (add_tdims_from_listobj(self, self->fits, tdimObj, nfields)) {
         status=99;
         goto create_table_cleanup;
     }
@@ -1979,7 +2024,7 @@ PyFITSObject_create_table_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
         if (extver > 0) {
 
             if (fits_update_key_lng(self->fits, "EXTVER", (LONGLONG) extver, NULL, &status)) {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, self);
                 goto create_table_cleanup;
             }
         }
@@ -1987,14 +2032,14 @@ PyFITSObject_create_table_hdu(struct PyFITSObject* self, PyObject* args, PyObjec
 
     if (nkeys > 0) {
         if (fits_set_hdrsize(self->fits, nkeys, &status) ) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             goto create_table_cleanup;
         }
     }
 
     // this does a full close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         goto create_table_cleanup;
     }
 
@@ -2035,12 +2080,12 @@ PyFITSObject_insert_col(struct PyFITSObject* self, PyObject* args, PyObject* kwd
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_insert_col(self->fits, colnum, ttype, tform, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2058,14 +2103,14 @@ PyFITSObject_insert_col(struct PyFITSObject* self, PyObject* args, PyObject* kwd
         free(tdim);
 
         if (status) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     }
 
     // this does a full close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2078,6 +2123,7 @@ PyFITSObject_insert_col(struct PyFITSObject* self, PyObject* args, PyObject* kwd
 // No error checking performed here
 static
 int write_string_column(
+        struct PyFITSObject* self,
         fitsfile *fits,  /* I - FITS file pointer                       */
         int  colnum,     /* I - number of column to write (1 = 1st col) */
         LONGLONG  firstrow,  /* I - first row to write (1 = 1st row)        */
@@ -2109,7 +2155,7 @@ int write_string_column(
     }
 
     if( fits_write_col_str(fits, colnum, firstrow, firstelem, nelem, strdata, status)) {
-        set_ioerr_string_from_status(*status);
+        set_ioerr_string_from_status(*status, self);
         free(strdata);
         return 1;
     }
@@ -2156,7 +2202,7 @@ PyFITSObject_write_column(struct PyFITSObject* self, PyObject* args, PyObject* k
     firstrow = (LONGLONG) firstrow_py;
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2194,25 +2240,25 @@ PyFITSObject_write_column(struct PyFITSObject* self, PyObject* args, PyObject* k
     if (fits_dtype == TSTRING) {
 
         // this is my wrapper for strings
-        if (write_string_column(self->fits, colnum, firstrow, firstelem, nelem, data, &status)) {
-            set_ioerr_string_from_status(status);
+        if (write_string_column(self, self->fits, colnum, firstrow, firstelem, nelem, data, &status)) {
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     } else if (fits_dtype == TBIT) {
         if (fits_write_col_bit(self->fits, colnum, firstrow, firstelem, nelem, data, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     } else {
         if( fits_write_col(self->fits, fits_dtype, colnum, firstrow, firstelem, nelem, data, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     }
 
     // this is a full file close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2263,7 +2309,7 @@ PyFITSObject_write_columns(struct PyFITSObject* self, PyObject* args, PyObject* 
     firstrow = (LONGLONG) firstrow_py;
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2354,14 +2400,15 @@ PyFITSObject_write_columns(struct PyFITSObject* self, PyObject* args, PyObject* 
         for (icol=0; icol<ncols; icol++) {
             data=PyArray_GETPTR1(array_ptrs[icol], irow);
             if (is_string[icol]) {
-                if (write_string_column(self->fits,
+                if (write_string_column(self,
+                                        self->fits,
                                         colnums[icol],
                                         thisrow,
                                         firstelem,
                                         nperrow[icol],
                                         (char*)data,
                                         &status)) {
-                    set_ioerr_string_from_status(status);
+                    set_ioerr_string_from_status(status, self);
                     goto _fitsio_pywrap_write_columns_bail;
                 }
 
@@ -2373,7 +2420,7 @@ PyFITSObject_write_columns(struct PyFITSObject* self, PyObject* args, PyObject* 
                                        nperrow[icol],
                                        data,
                                        &status)) {
-                    set_ioerr_string_from_status(status);
+                    set_ioerr_string_from_status(status, self);
                     goto _fitsio_pywrap_write_columns_bail;
                 }
             } else {
@@ -2385,7 +2432,7 @@ PyFITSObject_write_columns(struct PyFITSObject* self, PyObject* args, PyObject* 
                                    nperrow[icol],
                                    data,
                                    &status)) {
-                    set_ioerr_string_from_status(status);
+                    set_ioerr_string_from_status(status, self);
                     goto _fitsio_pywrap_write_columns_bail;
                 }
             }
@@ -2397,21 +2444,21 @@ PyFITSObject_write_columns(struct PyFITSObject* self, PyObject* args, PyObject* 
     if (fits_dtype == TSTRING) {
 
         // this is my wrapper for strings
-        if (write_string_column(self->fits, colnum, firstrow, firstelem, nelem, data, &status)) {
-            set_ioerr_string_from_status(status);
+        if (write_string_column(self, self->fits, colnum, firstrow, firstelem, nelem, data, &status)) {
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
 
     } else {
         if( fits_write_col(self->fits, fits_dtype, colnum, firstrow, firstelem, nelem, data, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     }
 
     // this is a full file close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     */
@@ -2489,6 +2536,7 @@ write_var_string_column_cleanup:
  */
 static
 int write_var_num_column(
+        struct PyFITSObject* self,
         fitsfile *fits,  /* I - FITS file pointer                       */
         int  colnum,     /* I - number of column to write (1 = 1st col) */
         LONGLONG  firstrow,  /* I - first row to write (1 = 1st row)        */
@@ -2542,7 +2590,7 @@ int write_var_num_column(
         Py_XDECREF(el_array);
 
         if(res > 0) {
-            set_ioerr_string_from_status(*status);
+            set_ioerr_string_from_status(*status, self);
             return 1;
         }
     }
@@ -2590,7 +2638,7 @@ PyFITSObject_write_var_column(struct PyFITSObject* self, PyObject* args, PyObjec
     firstrow = (LONGLONG) firstrow_py;
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2610,27 +2658,27 @@ PyFITSObject_write_var_column(struct PyFITSObject* self, PyObject* args, PyObjec
     // determine the fits dtype for this column.  We will use this to get data
     // from the array for writing
     if (fits_get_eqcoltypell(self->fits, colnum, &fits_dtype, NULL, NULL, &status) > 0) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_dtype == -TSTRING) {
         if (write_var_string_column(self->fits, colnum, firstrow, array, &status)) {
             if (status != 0) {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, self);
             }
             return NULL;
         }
     } else {
-        if (write_var_num_column(self->fits, colnum, firstrow, fits_dtype, array, &status)) {
-            set_ioerr_string_from_status(status);
+        if (write_var_num_column(self, self->fits, colnum, firstrow, fits_dtype, array, &status)) {
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     }
 
     // this is a full file close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2660,7 +2708,7 @@ PyFITSObject_write_record(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     // card not null terminated, so we copy everything.  GCC will
@@ -2668,13 +2716,13 @@ PyFITSObject_write_record(struct PyFITSObject* self, PyObject* args) {
     strncpy(card, cardin, FLEN_CARD);
 
     if (fits_write_record(self->fits, card, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2703,7 +2751,7 @@ PyFITSObject_write_string_key(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2712,13 +2760,13 @@ PyFITSObject_write_string_key(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_update_key_longstr(self->fits, keyname, value, comment, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2747,7 +2795,7 @@ PyFITSObject_write_double_key(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2756,13 +2804,13 @@ PyFITSObject_write_double_key(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_update_key_dbl(self->fits, keyname, value, decimals, comment, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2790,7 +2838,7 @@ PyFITSObject_write_long_long_key(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2799,13 +2847,13 @@ PyFITSObject_write_long_long_key(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_update_key_lng(self->fits, keyname, (LONGLONG) value, comment, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2832,7 +2880,7 @@ PyFITSObject_write_logical_key(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2841,13 +2889,13 @@ PyFITSObject_write_logical_key(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_update_key_log(self->fits, keyname, value, comment, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2872,18 +2920,18 @@ PyFITSObject_write_comment(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_write_comment(self->fits, comment, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2908,18 +2956,18 @@ PyFITSObject_write_history(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_write_history(self->fits, history, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -2975,18 +3023,18 @@ PyFITSObject_write_continue(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_write_continue(self->fits, value, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -3013,7 +3061,7 @@ PyFITSObject_write_undefined_key(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -3022,13 +3070,13 @@ PyFITSObject_write_undefined_key(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_update_key_null(self->fits, keyname, comment, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does not close and reopen
     if (fits_flush_buffer(self->fits, 0, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -3069,18 +3117,18 @@ PyFITSObject_insert_rows(struct PyFITSObject* self, PyObject* args, PyObject* kw
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_insert_rows(self->fits, firstrow, nrows, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does a full close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -3125,18 +3173,18 @@ PyFITSObject_delete_row_range(struct PyFITSObject* self, PyObject* args, PyObjec
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_delete_rows(self->fits, slice_start, nrows, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does a full close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -3178,18 +3226,18 @@ PyFITSObject_delete_rows(struct PyFITSObject* self, PyObject* args, PyObject* kw
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_delete_rowlistll(self->fits, rows, nrows, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     // this does a full close and reopen
     if (fits_flush_file(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -3461,7 +3509,7 @@ PyFITSObject_read_column(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -3479,7 +3527,7 @@ PyFITSObject_read_column(struct PyFITSObject* self, PyObject* args) {
 
     if (hdutype == ASCII_TBL) {
         if (read_ascii_column(self->fits, colnum, array, rows_array, sortind_array, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     } else {
@@ -3495,7 +3543,7 @@ PyFITSObject_read_column(struct PyFITSObject* self, PyObject* args) {
         }
 
         if (read_binary_column(self->fits, colnum, nrows, rows, sortind, data, stride, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
     }
@@ -3625,7 +3673,7 @@ PyFITSObject_read_var_column_as_list(struct PyFITSObject* self, PyObject* args) 
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -3641,7 +3689,7 @@ PyFITSObject_read_var_column_as_list(struct PyFITSObject* self, PyObject* args) 
     }
 
     if (fits_get_coltypell(self->fits, colnum, &fits_dtype, &repeat, &width, &status) > 0) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -3703,7 +3751,7 @@ read_var_column_cleanup:
         Py_XDECREF(tempObj);
         free_all_python_list(listObj);
         if (status != 0) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
         }
         return NULL;
     }
@@ -3845,7 +3893,7 @@ PyFITSObject_read_columns_as_rec(struct PyFITSObject* self, PyObject* args) {
 recread_columns_cleanup:
 
     if (status != 0) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     Py_RETURN_NONE;
@@ -4022,7 +4070,7 @@ PyFITSObject_read_columns_as_rec_byoffset(struct PyFITSObject* self, PyObject* a
 recread_columns_byoffset_cleanup:
 
     if (status != 0) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     Py_RETURN_NONE;
@@ -4158,7 +4206,7 @@ PyFITSObject_read_rows_as_rec(struct PyFITSObject* self, PyObject* args) {
 recread_byrow_cleanup:
 
     if (status != 0) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     Py_RETURN_NONE;
@@ -4229,7 +4277,7 @@ PyFITSObject_read_as_rec(struct PyFITSObject* self, PyObject* args) {
 recread_asrec_cleanup:
 
     if (status != 0) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     Py_RETURN_NONE;
@@ -4277,7 +4325,7 @@ PyFITSObject_read_image(struct PyFITSObject* self, PyObject* args) {
 
     if (fits_get_img_paramll(self->fits, maxdim, &datatype, &naxis,
                              naxes, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -4305,7 +4353,7 @@ PyFITSObject_read_image(struct PyFITSObject* self, PyObject* args) {
     }
     if (fits_read_pixll(self->fits, fits_read_dtype, firstpixels, size,
                         0, data, &anynul, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -4464,7 +4512,7 @@ PyFITSObject_read_image_slice(struct PyFITSObject* self, PyObject* args) {
 
     if (fits_read_subset(self->fits, fits_read_dtype, fpix, lpix, step,
                          0, data, &anynul, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         goto read_image_slice_cleanup;
     }
 
@@ -4540,12 +4588,12 @@ PyFITSObject_read_header(struct PyFITSObject* self, PyObject* args) {
         return NULL;
     }
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_get_hdrspace(self->fits, &nkeys, &morekeys, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -4555,7 +4603,7 @@ PyFITSObject_read_header(struct PyFITSObject* self, PyObject* args) {
         // the full card
         if (fits_read_record(self->fits, i+1, card, &status)) {
             Py_XDECREF(list);
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
 
@@ -4563,7 +4611,7 @@ PyFITSObject_read_header(struct PyFITSObject* self, PyObject* args) {
         // can eval in python
         if (fits_read_keyn(self->fits, i+1, keyname, value, scomment, &status)) {
             Py_XDECREF(list);
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, self);
             return NULL;
         }
 
@@ -4590,7 +4638,7 @@ PyFITSObject_read_header(struct PyFITSObject* self, PyObject* args) {
 
                 if (fits_read_key_longstr(self->fits, keyname, &longstr, comment, &status)) {
                     Py_XDECREF(list);
-                    set_ioerr_string_from_status(status);
+                    set_ioerr_string_from_status(status, self);
                     return NULL;
                 }
 
@@ -4750,16 +4798,16 @@ PyFITSObject_write_checksum(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_write_chksum(self->fits, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
     if (fits_get_chksum(self->fits, &datasum, &hdusum, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -4784,12 +4832,12 @@ PyFITSObject_verify_checksum(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
     if (fits_verify_chksum(self->fits, &dataok, &hduok, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -4835,7 +4883,7 @@ PyFITSObject_where(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_movabs_hdu(self->fits, hdunum, &hdutype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         return NULL;
     }
 
@@ -4846,7 +4894,7 @@ PyFITSObject_where(struct PyFITSObject* self, PyObject* args) {
     }
 
     if (fits_find_rows(self->fits, expression, firstrow, nrows, &ngood, row_status, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, self);
         goto where_function_cleanup;
     }
 
@@ -4919,7 +4967,7 @@ PyFITS_get_keytype(PyObject* self, PyObject* args) {
 
 
     if (fits_get_keytype(card, dtype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, NULL);
         return NULL;
     } else {
         return Py_BuildValue("s", dtype);
@@ -4941,7 +4989,7 @@ PyFITS_get_key_meta(PyObject* self, PyObject* args) {
     keyclass=fits_get_keyclass(card);
 
     if (fits_get_keytype(card, dtype, &status)) {
-        set_ioerr_string_from_status(status);
+        set_ioerr_string_from_status(status, NULL);
         return NULL;
     }
 
@@ -4986,11 +5034,11 @@ PyFITS_parse_card(PyObject* self, PyObject* args) {
     if (keyclass != TYP_COMM_KEY && keyclass != TYP_CONT_KEY) {
 
         if (fits_get_keyname(card, name, &keylen, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, NULL);
             goto bail;
         }
         if (fits_parse_value(card, value, comment, &status)) {
-            set_ioerr_string_from_status(status);
+            set_ioerr_string_from_status(status, NULL);
             goto bail;
         }
         if (fits_get_keytype(value, dtype, &status)) {
@@ -4999,7 +5047,7 @@ PyFITS_parse_card(PyObject* self, PyObject* args) {
                 is_undefined=1;
                 status=0;
             } else {
-                set_ioerr_string_from_status(status);
+                set_ioerr_string_from_status(status, NULL);
                 goto bail;
             }
         }